chore(SECURITY.md): Change expectations for initial report
The preparation of a nice security report can take time. We don't want security contributors to loose too much time preparing a report if the issue is already known. This PR changes the expectation for the initial report to state that it's ok for the initial contact to just send a summary of the issue.